sut/a2a-python/pyproject.toml
# Generated by A2A TCK code generator — do not edit by hand.

[project]
name = "a2a-tck-sut"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server,grpc,sqlite]>=0.3.0",
    "packaging",
    "starlette",
    "uvicorn",
]

[tool.uv.sources]
a2a-sdk = { path = "../../../a2a-python" }

[tool.ruff]
line-length = 130

[tool.ruff.lint.per-file-ignores]
"*.py" = ["I001", "F401", "PLR0912", "PLR0915"]
